backend: k8s: - support crawls with multiple wacz files, don't assume crawl complete after first wacz uploaded - if crawl is running and has wacz file, still show as running - k8s: allow configuring node selector for main pods (eg. nodeType=main) and for crawlers (eg. nodeType=crawling) - profiles: support uploading to alternate storage specified via 'shared_profile_storage' value is set - misc fixes for profiles

+++ backend/k8sman.py
@@ -172,7 +172,12 @@ class K8SManager:
             )
 
     async def add_crawl_config(
-        self, crawlconfig, storage, run_now, out_filename, profile_filename
+        self,
+        crawlconfig,
+        storage,
+        run_now,
+        out_filename,
+        profile_filename,
     ):
         """add new crawl as cron job, store crawl config in configmap"""
         cid = str(crawlconfig.id)
@@ -343,7 +348,7 @@ class K8SManager:
             return None, None
 
         manual = job.metadata.annotations.get("btrix.run.manual") == "1"
-        if manual and not self.no_delete_jobs:
+        if manual and not self.no_delete_jobs and crawlcomplete.completed:
             self.loop.create_task(self._delete_job(job.metadata.name))
 
         crawl = self._make_crawl_for_job(
@@ -389,12 +394,14 @@ class K8SManager:
             endpoint_url = self._secret_data(storage_secret, "STORE_ENDPOINT_URL")
             access_key = self._secret_data(storage_secret, "STORE_ACCESS_KEY")
             secret_key = self._secret_data(storage_secret, "STORE_SECRET_KEY")
+            region = self._secret_data(storage_secret, "STORE_REGION") or ""
 
             self._default_storages[name] = S3Storage(
                 access_key=access_key,
                 secret_key=secret_key,
                 endpoint_url=endpoint_url,
                 access_endpoint_url=access_endpoint_url,
+                region=region,
             )
 
         return self._default_storages[name]
@@ -542,17 +549,19 @@ class K8SManager:
             return True
 
     async def run_profile_browser(
-        self, userid, aid, storage, command, baseprofile=None
+        self, userid, aid, command, storage=None, storage_name=None, baseprofile=None
     ):
         """run browser for profile creation """
-        # Configure Annotations + Labels
-        if storage.type == "default":
+
+        # if default storage, use name and path + profiles/
+        if storage:
             storage_name = storage.name
-            storage_path = storage.path
+            storage_path = storage.path + "profiles/"
+        # otherwise, use storage name and existing path from secret
         else:
-            storage_name = aid
             storage_path = ""
 
+        # Configure Annotations + Labels
         labels = {
             "btrix.user": userid,
             "btrix.archive": aid,
@@ -560,7 +569,7 @@ class K8SManager:
         }
 
         if baseprofile:
-            labels["btrix.baseprofile"] = baseprofile
+            labels["btrix.baseprofile"] = str(baseprofile)
 
         await self.check_storage(storage_name)
 
@@ -835,6 +844,7 @@ class K8SManager:
                 "template": {
                     "metadata": {"labels": labels},
                     "spec": {
+                        "nodeSelector": {"nodeType": "crawling"},
                         "containers": [
                             {
                                 "name": "crawler",
@@ -891,7 +901,7 @@ class K8SManager:
                             },
                             self.crawl_volume,
                         ],
-                        "restartPolicy": "Never",
+                        "restartPolicy": "OnFailure",
                         "terminationGracePeriodSeconds": self.grace_period,
                     },
                 },
